refactor: order event producer DI 패턴으로 변경
